requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
coreapi==2.3.3
psycopg2==2.9.9


